# Generated_Server_Python/openapi_server/openapi/openapi.yaml
openapi: 3.0.0
info:
  contact:
    email: office@deepviss.org
  description: DeepVISS (Deep Vision Interoperability Specification Standard) allows
    several computer vision solutions to produce, consume and exchange events in the
    same format.
  license:
    name: Apache 2.0
    url: https://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: https://deepviss.org
  title: DeepVISS TAG
  version: 1.2.0
...
servers:
- url: /
...
paths:
  /challenge:
    post:
      operationId: create_challenge
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Challenge'
          application/xml:
            schema:
              $ref: '#/components/schemas/Challenge'
        description: The challenge object to be created
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/definitions/Challenge'
          description: successful operation
        "400":
          description: Invalid status value
      summary: Create a new challenge.
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenge/assign:
    post:
      operationId: assign_challenge
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AssignUsersRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/AssignUsersRequest'
        description: The users to be assigned to a challenge
        required: true
      responses:
        "200":
          content: {}
          description: successful operation
        "400":
          description: Invalid status value
      summary: Assign a private challenge.
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenge/broker/list:
    get:
      description: Multiple status values can be provided with comma separated strings
      operationId: get_challenge_list_broker
      parameters:
      - explode: true
        in: query
        name: challengeName
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: accessType
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: liked
        required: false
        schema:
          type: boolean
        style: form
      - explode: true
        in: query
        name: pageSize
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - explode: true
        in: query
        name: pageNumber
        required: false
        schema:
          format: int32
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/definitions/Challenge'
                type: array
          description: successful operation
        "400":
          description: Invalid status value
      summary: Retrieve list of challenges
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenge/collaborator/list:
    get:
      description: Multiple status values can be provided with comma separated strings
      operationId: get_challenge_list_collaborator
      parameters:
      - explode: true
        in: query
        name: challengeName
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: accessType
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: liked
        required: false
        schema:
          type: boolean
        style: form
      - explode: true
        in: query
        name: pageSize
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - explode: true
        in: query
        name: pageNumber
        required: false
        schema:
          format: int32
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/definitions/Challenge'
                type: array
          description: successful operation
        "400":
          description: Invalid status value
      summary: Retrieve list of challenges
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenge/cost:
    post:
      description: estimage challenge budget
      operationId: get_challenge_cost
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChallengeBudgetRequest'
          application/xml:
            schema:
              $ref: '#/components/schemas/ChallengeBudgetRequest'
        description: The challenge options
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeBudgetResponse'
          description: successful operation
        "400":
          description: Invalid status value
      summary: get estimated cost for challenge
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenge/supervisor/list:
    get:
      description: Multiple status values can be provided with comma separated strings
      operationId: get_challenge_list_supervisor
      parameters:
      - explode: true
        in: query
        name: challengeName
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: accessType
        required: false
        schema:
          type: string
        style: form
      - explode: true
        in: query
        name: liked
        required: false
        schema:
          type: boolean
        style: form
      - explode: true
        in: query
        name: pageSize
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - explode: true
        in: query
        name: pageNumber
        required: false
        schema:
          format: int32
          type: integer
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/definitions/Challenge'
                type: array
          description: successful operation
        "400":
          description: Invalid status value
      summary: Retrieve list of challenges
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenge/{challengeId}:
    get:
      description: Multiple status values can be provided with comma separated strings
      operationId: get_challenge
      parameters:
      - description: The identifier of the dataset
        explode: false
        in: path
        name: challengeId
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: accessToken
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/definitions/Challenge'
          description: successful operation
        "400":
          description: Invalid status value
      summary: Retrieve list of challenges
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
    patch:
      operationId: patch_challenge
      parameters:
      - description: The identifier of the dataset
        explode: false
        in: path
        name: challengeId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchChallenge'
          application/xml:
            schema:
              $ref: '#/components/schemas/PatchChallenge'
        description: The new challenge object
        required: true
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/definitions/Challenge'
          description: successful operation
        "400":
          description: Invalid status value
      summary: Update challenge.
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenge/{challengeId}/archive/list:
    get:
      description: Multiple status values can be provided with comma separated strings
      operationId: archived_shard_list
      parameters:
      - description: The identifier of the challenge
        explode: false
        in: path
        name: challengeId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/definitions/ArchivedShard'
                type: array
          description: successful operation
        "400":
          description: Invalid status value
      summary: Get archive list from challenges
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenge/{challengeId}/archive/{shard}:
    get:
      description: Multiple status values can be provided with comma separated strings
      operationId: archive_shard
      parameters:
      - description: The identifier of the challenge
        explode: false
        in: path
        name: challengeId
        required: true
        schema:
          type: string
        style: simple
      - description: The identifier of the shard
        explode: false
        in: path
        name: shard
        required: true
        schema:
          type: integer
        style: simple
      - explode: true
        in: query
        name: force
        required: false
        schema:
          type: boolean
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/definitions/ArchivedShard'
          description: successful operation
        "400":
          description: Invalid status value
      summary: Archive shard from challenges
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenge/{challengeId}/assign-tenant/{tenantId}:
    post:
      operationId: assign_tenant_to_challenge
      parameters:
      - explode: false
        in: path
        name: challengeId
        required: true
        schema:
          type: string
        style: simple
      - explode: false
        in: path
        name: tenantId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content: {}
          description: successful operation
        "400":
          description: Invalid status value
      summary: Star a public challenge.
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenge/{challengeId}/download/{shard}:
    get:
      description: Multiple status values can be provided with comma separated strings
      operationId: download_challenge
      parameters:
      - description: The identifier of the challenge
        explode: false
        in: path
        name: challengeId
        required: true
        schema:
          type: string
        style: simple
      - description: The identifier of the shard
        explode: false
        in: path
        name: shard
        required: true
        schema:
          type: integer
        style: simple
      - explode: true
        in: query
        name: X-Auth-Credential
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          description: successful operation
        "400":
          description: Invalid status value
      summary: Download tagged items from challenges
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenge/{challengeId}/force-merge:
    get:
      description: Multiple status values can be provided with comma separated strings
      operationId: force_merge_tagged_media_item
      parameters:
      - description: The identifier of the dataset
        explode: false
        in: path
        name: challengeId
        required: true
        schema:
          type: string
        style: simple
      - description: The identifier of the tagged media item
        explode: true
        in: query
        name: challengeName
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/definitions/Challenge'
          description: successful operation
        "400":
          description: Invalid status value
      summary: force merge
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenge/{challengeId}/reject:
    post:
      operationId: reject_challenge
      parameters:
      - explode: false
        in: path
        name: challengeId
        required: true
        schema:
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              type: string
        description: The challenge object to be created
        required: true
      responses:
        "200":
          content: {}
          description: successful operation
        "400":
          description: Invalid status value
      summary: Star a public challenge.
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenge/{challengeId}/star:
    post:
      operationId: star_challenge
      parameters:
      - explode: false
        in: path
        name: challengeId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content: {}
          description: successful operation
        "400":
          description: Invalid status value
      summary: Star a public challenge.
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenge/{challengeId}/statistics:
    get:
      operationId: statistics_challenge
      parameters:
      - explode: false
        in: path
        name: challengeId
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChallengeStatistics'
            application/xml:
              schema:
                $ref: '#/components/schemas/ChallengeStatistics'
          description: successful operation
        "400":
          description: Invalid status value
      summary: Get challenge statistics
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
  /challenge/{challengeId}/tagged-item/list:
    get:
      description: Multiple status values can be provided with comma separated strings
      operationId: get_tagged_items
      parameters:
      - description: The identifier of the dataset
        explode: false
        in: path
        name: challengeId
        required: true
        schema:
          type: string
        style: simple
      - explode: true
        in: query
        name: pageSize
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - explode: true
        in: query
        name: pageNumber
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - explode: true
        in: query
        name: stage
        required: false
        schema:
          format: int32
          type: integer
        style: form
      - explode: true
        in: query
        name: accessToken
        required: false
        schema:
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: '#/definitions/TaggedMediaItem'
                type: array
          description: successful operation
        "400":
          description: Invalid status value
      summary: Retrieve list tagged media items
      tags:
      - challenge
      x-openapi-router-controller: openapi_server.controllers.challenge_controller
...
components:
  schemas:
...
    Challenge:
      example:
        completionPercentageForCollaborator: 70
        dispatchPolicyId: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
        supplierTenantId: Dogs
        edgeTaxonomy: is a, has a
        initialNumberOfShards: 4
        demanderTenantId: Dogs
        description: Tag the cats
        resolution: Tag the cats
        nodeTaxonomy: consumer,overdue
        completionPercentages:
        - stages:
          - percentage: 100
            stageNumber: 1
          - percentage: 100
            stageNumber: 1
          shard: 1
        - stages:
          - percentage: 100
            stageNumber: 1
          - percentage: 100
            stageNumber: 1
          shard: 1
        creationTimestamp: 2000-01-23T04:56:07.000+00:00
        datasetId: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
        budget: 3.14
        challengeAccessToken: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
        datasetMetadata:
          mediaItems: 1000
          numberOfShards: 4
          name: Dogs
        completionPercentage: 70
        challengeId: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
        name: Dogs
        expiration: 4
        evaluationPolicyId: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
        stared: true
        estimatedTagsPerMediaItem: 4
      properties:
        challengeId:
          description: 'Alpha-numeric, unique id of challenge. '
          example: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
          type: string
        datasetId:
          description: Alpha-numeric, unique id of dataset
          example: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
          type: string
        evaluationPolicyId:
          description: Alpha-numeric, unique id of evaluation policy
          example: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
          type: string
        dispatchPolicyId:
          description: Alpha-numeric, unique id of dispatch policy
          example: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
          type: string
        creationTimestamp:
          description: Date and time expressed according to ISO 8601 (e.g. 2018-06-24T23:10:28+03:00)
          format: date-time
          type: string
        status:
          $ref: '#/components/schemas/ChallengeStatusEnum'
        accessType:
          $ref: '#/components/schemas/ChallengeAccessTypeEnum'
        tenantRole:
          $ref: '#/components/schemas/ChallengeTenantRoleEnum'
        name:
          description: '[TBD] Alpha-numeric, name of dataset'
          example: Dogs
          type: string
        completionPercentage:
          description: completion percentage of a challenge
          example: 70
          type: integer
        completionPercentageForCollaborator:
          description: completion percentage of a challenge for a collaborator
          example: 70
          type: integer
        description:
          description: '[TBD] Alpha-numeric, name of dataset'
          example: Tag the cats
          type: string
        budget:
          description: budget allocated for challenge
          example: 3.14
          format: double
          type: number
        stared:
          description: '[TBD] Alpha-numeric, name of dataset'
          example: true
          type: boolean
        supplierTenantId:
          description: '[TBD] Alpha-numeric, name of supplierTenantId'
          example: Dogs
          type: string
        resolution:
          description: '[TBD] Alpha-numeric, name of dataset'
          example: Tag the cats
          type: string
        demanderTenantId:
          description: '[TBD] Alpha-numeric, name of demanderTenantId'
          example: Dogs
          type: string
        initialNumberOfShards:
          description: shards in dataset
          example: 4
          type: integer
        datasetMetadata:
          $ref: '#/components/schemas/Challenge_datasetMetadata'
        tagType:
          $ref: '#/components/schemas/ShapeTypeEnum'
        estimatedTagsPerMediaItem:
          description: estimated Tags Per Media Item
          example: 4
          format: double
          type: number
        nodeTaxonomy:
          description: taxonomy list for nodeType
          example: consumer,overdue
          items:
            type: string
          type: array
        edgeTaxonomy:
          description: taxonomy list for edgeType
          example: is a, has a
          items:
            type: string
          type: array
        completionPercentages:
          items:
            $ref: '#/components/schemas/ChallengeCompletionPercentage'
          type: array
        expiration:
          description: hours until the challenge is expired
          example: 4
          type: integer
        challengeAccessToken:
          description: Token for accessing the challenge
          example: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
          type: string
      required:
      - accessType
      type: object
    PatchChallenge:
      example:
        dispatchPolicyId: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
        name: Dogs
        description: Tag the cats
        evaluationPolicyId: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
        budget: 3.14
      properties:
        evaluationPolicyId:
          description: Alpha-numeric, unique id of evaluation policy
          example: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
          type: string
        dispatchPolicyId:
          description: Alpha-numeric, unique id of dispatch policy
          example: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
          type: string
        accessType:
          $ref: '#/components/schemas/ChallengeAccessTypeEnum'
        name:
          description: '[TBD] Alpha-numeric, name of dataset'
          example: Dogs
          type: string
        description:
          description: '[TBD] Alpha-numeric, name of dataset'
          example: Tag the cats
          type: string
        budget:
          description: budget allocated for challenge
          example: 3.14
          format: double
          type: number
      type: object
    ChallengeAccessTypeEnum:
      enum:
      - public
      - private
      - spot
      - confidential
      type: string
    ChallengeStatusEnum:
      enum:
      - pending
      - inProgress
      - completed
      - rejected
      type: string
    ChallengeTenantRoleEnum:
      enum:
      - SUPPLIER
      - BROKER
      - DEMANDER
      type: string
    ChallengeBudgetRequest:
      example:
        dispatchPolicyId: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
        datasetId: XAFKE80768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
        evaluationPolicyId: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
        estimatedTagsPerMediaItem: 4
      properties:
        datasetId:
          description: '[TBD] Alpha-numeric, unique id of dataset'
          example: XAFKE80768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
          type: string
        evaluationPolicyId:
          description: Alpha-numeric, unique id of evaluation policy
          example: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
          type: string
        dispatchPolicyId:
          description: Alpha-numeric, unique id of dispatch policy
          example: DFK768EF053898993DD196397EDFDDFBC81751818B7FD1300124455B07E91CB289F87791D78064ECC93754F19B13D419489F162A150A22DD814CKAF0E
          type: string
        tagType:
          $ref: '#/components/schemas/ShapeTypeEnum'
        estimatedTagsPerMediaItem:
          description: estimated Tags Per Media Item
          example: 4
          format: double
          type: number
      type: object
    ChallengeBudgetResponse:
      example:
        estimatedBudget: 3.14
      properties:
        estimatedBudget:
          description: estimated budget for challenge
          example: 3.14
          format: double
          type: number
      type: object
...
    AssignUsersRequest:
      description: assign users dto
      example:
        challengeId: challengeId
        avatarIds:
        - avatarIds
        - avatarIds
      properties:
        challengeId:
          description: Unique ID of challenge
          type: string
        avatarIds:
          items:
            type: string
          type: array
      type: object
...
    ShapeTypeEnum:
      enum:
      - 2D_rectangle
      - 2D_polygon
      - 2D_rectangle_text_block
      - 2D_polygon_text_block
      - 2D_timebound_polygon
      type: string
...
    Challenge_datasetMetadata:
      example:
        mediaItems: 1000
        numberOfShards: 4
        name: Dogs
      properties:
        name:
          description: '[TBD] Alpha-numeric, name of dataset'
          example: Dogs
          type: string
        mediaItems:
          description: estimated Tags Per Media Item
          example: 1000
          type: integer
        numberOfShards:
          description: shards in dataset
          example: 4
          type: integer
...